Split code into function
