+++ src/commands/runtime.ts
@@ -1,6 +1,7 @@
+import * as vscode from 'vscode'
 
+import {selectScript} from "./helper/ui"
 
-import * as vscode from 'vscode';
 
 export const openRepl = () => {
 	const terminal = vscode.window.createTerminal({ name: 'Arturo REPL' })
@@ -9,24 +10,9 @@ export const openRepl = () => {
 };
 
 export const runFile = async () => {
-	const uris = await vscode.workspace.findFiles('**/*.art')
-	if (!uris || uris.length === 0) {
-		vscode.window.showInformationMessage('No Arturo file found in workspace.')
-		return
-	}
+	const file = await selectScript()
+    if (!file) return
 
-	const items = uris.map(u => ({
-		label: vscode.workspace.asRelativePath(u),
-		uri: u
-	})) as Array<vscode.QuickPickItem & { uri: vscode.Uri }>;
-
-	const pick = await vscode.window.showQuickPick(items, { 
-        placeHolder: 'Select a file to run with Arturo' 
-    })
-
-	if (!pick) return
-
-	const file = (pick as any).uri.fsPath as string
 	const terminal = vscode.window.createTerminal({ name: 'Arturo Run' })
 	terminal.show(true)
 	terminal.sendText(`arturo "${file}"`)
